requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
packaging==24.0
playwright==1.44.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.2.1
requests==2.32.3
tomli==2.0.1
typing_extensions==4.12.0
urllib3==2.2.1
